Moved from blueprintjs to material-ui
It seems like it's easier to. Plus I have some experience with it.

Added the nice side bar, no toggling yet though, since it needs to go
through redux.

Also added a school list that you can collapse.

Finally moved to a select filter for the schools rather than a bunch of
buttons.